requirements.txt
asgiref==3.7.2
attrs==23.2.0
Automat==22.10.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
build==1.0.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.2
cssselect==1.2.0
dbus-python==1.3.2
distlib==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.9
django-bootstrap-v5==1.0.11
django-bootstrap5==23.4
django-environ==0.11.2
djangorestframework==3.14.0
duplicity==2.2.2
fasteners==0.19
filelock==3.13.1
fontawesomefree==6.5.1
future==0.18.3
gunicorn==21.2.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jaraco.classes==3.3.0
jeepney==0.8.0
jmespath==1.0.1
keyring==24.3.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lockfile==0.12.2
lxml==5.1.0
macaroonbakery==1.3.4
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.4
monotonic==1.6
more-itertools==10.2.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.47
packaging==23.2
paramiko==3.4.0
parsel==1.8.1
pexpect==4.9.0
pillow==10.2.0
pip-autoremove==0.10.0
pip-review==1.3.0
pip-tools==7.3.0
pipenv==2023.11.17
platformdirs==4.2.0
priority==2.0.0
Protego==0.3.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.25.1
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.17.2
PyGObject==3.46.0
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyRFC3339==1.1
python-dateutil==2.8.2
python-debian==0.1.49
python-decouple==3.8
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
queuelib==1.6.2
reportlab==4.1.0
requests==2.31.0
requests-file==2.0.0
SecretStorage==3.3.3
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
supervisor==4.2.5
systemd-python==235
tldextract==5.1.1
tomli==2.0.1
Twisted==23.10.0
typing_extensions==4.9.0
urllib3==2.2.1
virtualenv==20.25.0
virtualenv-clone==0.5.7
w3lib==2.1.2
wadllib==1.3.6
whitenoise==6.6.0
xdg==6.0.0
zipp==3.17.0
zope.interface==6.1
